Правки по PR
