+++ src/test/java/com/example/CatTest.java
@@ -2,15 +2,20 @@ package com.example;
 
 import org.junit.Assert;
 import org.junit.Test;
+import org.junit.runner.RunWith;
 import org.mockito.Mock;
+import org.mockito.Mockito;
+import org.mockito.junit.MockitoJUnitRunner;
 
 import java.util.List;
 
 import static org.junit.Assert.*;
 
+@RunWith(MockitoJUnitRunner.class)
 public class CatTest {
     @Mock
     Feline feline;
+
 
     @Test
     public void getSoundTest() {
@@ -21,9 +26,10 @@ public class CatTest {
 
     @Test
     public void getFoodTest() throws Exception {
-        Cat cat = new Cat(new Feline());
-        List<String> actualFoodList = cat.getFood();
-        List<String> expectedFoodList = List.of("Животные", "Птицы", "Рыба");
-        Assert.assertEquals(expectedFoodList, actualFoodList);
+        Cat cat = new Cat(feline);
+        Mockito.when(feline.eatMeat()).thenReturn(List.of("Животные", "Птицы", "Рыба"));
+        List<String> expectedList = List.of("Животные", "Птицы", "Рыба");
+        List<String> actualList = cat.getFood();
+        assertEquals(expectedList, actualList);
     }
 }
