+++ src/test/java/com/example/LionTest.java
@@ -1,11 +1,7 @@
 package com.example;
 
 import org.junit.Before;
-import org.junit.Rule;
 import org.junit.Test;
-import org.junit.rules.ExpectedException;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 import org.mockito.Mock;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
@@ -14,23 +10,7 @@ import java.util.List;
 
 import static org.junit.Assert.*;
 
-@RunWith(Parameterized.class)
 public class LionTest {
-    private final String checkedSex;
-    private final boolean expectedSex;
-
-    public LionTest(String checkedSex, boolean expectedSex) {
-        this.checkedSex = checkedSex;
-        this.expectedSex = expectedSex;
-    }
-
-    @Parameterized.Parameters
-    public static Object[][] getSexData() {
-        return new Object[][]{
-                {"Самец", true},
-                {"Самка", false},
-        };
-    }
 
     @Before
     public void init() {
@@ -48,22 +28,11 @@ public class LionTest {
         assertEquals(1, actualKittensCount);
     }
 
-
     @Test
-    public void doesHaveManeTest() throws Exception {
-        Lion lion = new Lion(checkedSex, feline);
-        boolean actualSex = lion.doesHaveMane();
-        assertEquals(expectedSex, actualSex);
-    }
-
-    @Rule
-    public ExpectedException thrown = ExpectedException.none();
-
-    @Test
-    public void shouldCreatingLionThrowExceptionTest() throws Exception {
-        thrown.expect(Exception.class);
-        thrown.expectMessage("Используйте допустимые значения пола животного - самец или самка");
-        new Lion("котёнок", feline);
+    public void shouldCreatingLionThrowExceptionTest() {
+        Exception exception = assertThrows(Exception.class,
+                () -> new Lion("котёнок", feline));
+        assertTrue(exception.getMessage().contains("Используйте допустимые значения пола животного - самец или самка"));
     }
 
 
@@ -80,6 +49,7 @@ public class LionTest {
         Lion lion = new Lion("Самец", feline);
         Mockito.when(feline.getFood("Хищник")).thenThrow(Exception.class);
         lion.getFood();
+        assertThrows(Exception.class, () -> lion.getFood());
 
     }
